frontend/pyproject.toml
[project]
name = "frontend"
version = "1.0.0"
description = "Flask frontend for Task Management System"
requires-python = ">=3.14"
dependencies = [
    "Flask==3.0.0",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "Werkzeug==3.0.1",
    "granian>=2.5.6",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]
